aiCall.js 에서 aiRedering.js로 stringSplit과 notDollar 함수 모듈화
