+++ JS/aiQuestion/aiCall.js
@@ -1,4 +1,4 @@
-import {talkRendering} from './aiRendering.js';
+import {talkRendering, stringSplit} from './aiRendering.js';
 
 const API_KEY = 'AIzaSyAnx5WFFsMBgfx8dmdEruWmT5888F5TJCI';
 const url = `https://generativelanguage.googleapis.com/v1beta/models/gemini-2.0-flash:generateContent?key=${API_KEY}`;
@@ -47,71 +47,7 @@ async function aiCall() {
     const endIndex = problemJsonString.lastIndexOf('}');
     problemJsonString = problemJsonString.substring(startIndex, endIndex + 1);
   }
-  stringSplit(problemJsonString);
-
-
-}
-
-function stringSplit(jsonString) {
-  const obj = JSON.parse(jsonString);// 문자열 → 객체
-
-
-  let resultText = notDollar(obj);
-  console.log("not$", resultText);
-  // ":"를 기준으로 알아서 key와 value로 나눠줌
-  Object.entries(obj).forEach(([key, value]) => {
-    const label = {
-      problemType: '문제 유형',
-      problemLevel: '문제 레벨',
-      problem: '문제',
-      problemOneLine: '문제 한줄 설명',
-      solvingOrder: ['문제 풀이'],
-      answer: '문제의 답',
-      tip: '비슷한 유형에서의 풀이법'
-    }[key] || key;
-
-    // value가 배열일 경우 예쁘게 줄바꿈 처리
-    if (Array.isArray(value)) {
-      resultText += `${label}:\n${value.map(v => ` - ${v}`).join('\n\n')}\n\n`;
-    } else {
-      resultText += `${label}: ${value}\n\n`;
-    }
-  });
-
-  console.log('str', resultText);
-
-  // aiTalkRendering(resultText)
-  talkRendering('ai', resultText);
-}
-
-//======달러표시 지우기===///
-function notDollar(jsonString) {
-  function removeDollarSigns(text) {
-    // 백슬래시로 이스케이프된 $도 함께 제거하기 위해 \\$ 패턴도 고려
-    return text.replace(/\\?\$/g, '');
-  }
-
-  // 객체 내의 모든 문자열 값에서 $ 기호 제거
-  function processObject(data) {
-    for (const key in data) {
-      if (typeof data[key] === 'string') {
-        data[key] = removeDollarSigns(data[key]);
-      } else if (Array.isArray(data[key])) {
-        data[key] = data[key].map(item => {
-          if (typeof item === 'string') {
-            return removeDollarSigns(item);
-          }
-          return item;
-        });
-      } else if (typeof data[key] === 'object' && data[key] !== null) {
-        processObject(data[key]); // 중첩된 객체 처리 (현재 JSON 구조에서는 필요 없지만, 일반적인 처리 방식)
-      }
-    }
-    return data;
-  }
-  const processedObj = processObject(jsonString);
-  // 수정된 객체를 다시 JSON 문자열로 변환 (필요하다면)
-  const cleanedJsonString = JSON.stringify(processedObj, null, 2);
+  talkRendering('ai', stringSplit(problemJsonString));
 }
 
 //===========프롬프트========//
